% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{getWintimeIntegral}
\alias{getWintimeIntegral}
\title{Helper functions for package functions}
\usage{
getWintimeIntegral(m, etimes, time0, time1, delta0, delta1)
}
\arguments{
\item{m}{The number of events in the hierarchy.}

\item{etimes}{A sorted vector of event times (days) (returned from wintime::setEventTimes()).}

\item{time0}{A vector containing the control person's event times (days).}

\item{time1}{A vector containing the treatment person's event times (days).}

\item{delta0}{A vector containing the control person's event indicators.}

\item{delta1}{A vector containing the treatment person's event indicators.}
}
\value{
The win time difference integral.
}
\description{
Win time difference
}
\details{
This function calculates the win time difference integral for a single pair. This function is used in all pairwise win time methods.
}
