% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{getWintimeIntegral_rest}
\alias{getWintimeIntegral_rest}
\title{Win time difference with time restriction}
\usage{
getWintimeIntegral_rest(
  m,
  etimes,
  time0,
  time1,
  delta0,
  delta1,
  time_restriction
)
}
\arguments{
\item{m}{The number of events in the hierarchy.}

\item{etimes}{A sorted vector of event times (days) (returned from wintime::setEventTimes()).}

\item{time0}{A vector containing the control person's event times (days).}

\item{time1}{A vector containing the treatment person's event times (days).}

\item{delta0}{A vector containing the control person's event indicators.}

\item{delta1}{A vector containing the treatment person's event indicators.}

\item{time_restriction}{The time restriction (days) for calculation.}
}
\value{
A list of the win time difference integral and its components.
}
\description{
This function calculates the win time difference integral for a single pair with truncation at time_restriction.
This function is used in all pairwise win time methods.
}
