% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{setKM}
\alias{setKM}
\title{Set event times and indicators used in the Kaplan-Meier survival curve calculation}
\usage{
setKM(n, m, time, delta)
}
\arguments{
\item{n}{The total number of trial participants.}

\item{m}{The number of events in the hierarchy.}

\item{time}{The row reversal of the Time matrix (days) (created inside wintime::km()).}

\item{delta}{The row reversal of the Delta matrix (created inside wintime::km()).}
}
\value{
A list containing the event time matrix and the event indicator matrix used in wintime::km().
}
\description{
This function creates the time_km and delta_km matrices used for wintime::km().
}
